Rename domain module as entity
