+++ packster-core/src/entity.rs
