+++ packster-core/src/port.rs
@@ -1,7 +1,7 @@
 use std::{path::Path, io::{Read, Write}};
 use serde::{de::DeserializeOwned, ser::Serialize};
 
-use crate::{Result, path::{Absolute, NormalizedPathBuf}, Error, domain::Checksum};
+use crate::{Result, path::{Absolute, NormalizedPathBuf}, Error, entity::Checksum};
 
 pub trait PathExt {
     fn is_ancestor_of<P: AsRef<Path>>(&self, child_path: P) -> bool;
